+++ src/dir.rs
@@ -731,6 +731,14 @@ impl Dir {
         if name.0 == b"." || name.0 == b".." {
             return Err(Ext4Error::DotEntry);
         }
+
+        let linked_inode =
+            get_dir_entry_inode_by_name(&self.fs, &self.inode, name).await?;
+        assert_eq!(
+            linked_inode.index, inode.index,
+            "unlink called with inode that does not match directory entry"
+        );
+
         let old = inode.links_count();
         inode.set_links_count(old.saturating_sub(1));
         inode.write(&self.fs).await?;